Fix integer overflow
The energyCost parameter of the SimpleFusionRecipe constructor is multiplied by the number of pedestals used. A cost of 1,000,000,000 will cause an integer overflow since the total cost is stored as signed 32bit integer. If the total crafting cost was supposed to be 1bn RF, 128,000,000 would be the correct amount in the function call. Alternatively change the amount to 256m to match the most expensive chaotic recipe in DE with 2bn RF.